backend/requirements.txt
Django==6.0.3
djangorestframework==3.17.1
django-cors-headers==4.6.0
Pillow==11.1.0
python-dotenv==1.1.0
google-generativeai==0.8.5
tensorflow==2.19.0
numpy>=1.26.0
scikit-learn>=1.4.0
qrcode[pil]==8.0
